add flag ToggleSyncLocalSave to allow local saves to sync to remote server if they are newer.  helpful if disconnected from server

+++ Source/Client/Managers/SaveManager.cs
@@ -55,21 +55,22 @@ namespace GameClient
                 File.WriteAllBytes(serverSaveFilePath, save);
                 File.Delete(tempSaveFilePath);
 
-                Logger.Message("Comparing remote vs local save (if exists)");
+                if(fileTransferData.instructions != (int)SaveMode.Strict) { 
+                    Logger.Message("Comparing remote vs local save (if exists)");
 
-                if (float.Parse(GetRealPlayTimeInteractingFromSave(serverSaveFilePath)) >=
-                float.Parse(GetRealPlayTimeInteractingFromSave(saveFilePath)))
-                {
-                    Logger.Message("Loading remote save");
-                    File.Delete(saveFilePath);
-                    File.Move(serverSaveFilePath, saveFilePath);
+                    if (float.Parse(GetRealPlayTimeInteractingFromSave(serverSaveFilePath)) >=
+                    float.Parse(GetRealPlayTimeInteractingFromSave(saveFilePath)))
+                    {
+                        Logger.Message("Loading remote save");
+                        File.Delete(saveFilePath);
+                        File.Move(serverSaveFilePath, saveFilePath);
+                    }
+                    else
+                    {
+                        Logger.Message("Loading local save");
+                        File.Delete(serverSaveFilePath);
+                    }
                 }
-                else
-                {
-                    Logger.Message("Loading local save");
-                    File.Delete(serverSaveFilePath);
-                }
-
                 GameDataSaveLoader.LoadGame(customSaveName);
                 return;
             }
